refactor: move context to internal
