+++ flow/internal/context.go
@@ -1,4 +1,4 @@
-package shared
+package internal
 
 import (
 	"context"
